Use a symlink instead of duplicating files
Make `examples/style/tar23.png` a symlink to `examples/input/in23.png` instead of duplicating the file in the repo. This not only saves disk space — it also reduces confusion by making it explicit/obvious to anyone browsing through the repo that these files are intended to be identical.

Ref. #6 & #8.